requirements.txt
streamlit
pandas
google_cloud_vision
gspread
oauth2client
thefuzz[speedup]
Pillow